Resolve merge changes, and some namespacing
