--- pems-server/src/main/java/org/maroubra/pemsserver/bindings/ActuatorBindings.java
@@ -1,10 +1,0 @@
-package org.maroubra.pemsserver.bindings;
-
-import org.glassfish.hk2.utilities.binding.AbstractBinder;
-
-public class ActuatorBindings extends AbstractBinder {
-    @Override
-    protected void configure() {
-        // TODO
-    }
-}
