+++ pems-server/src/main/java/org/maroubra/pemsserver/bindings/ServerBindings.java
@@ -1,6 +1,8 @@
 package org.maroubra.pemsserver.bindings;
 
 import com.google.inject.AbstractModule;
+import org.maroubra.pemsserver.control.ControlService;
+import org.maroubra.pemsserver.control.ControlServiceImpl;
 import org.maroubra.pemsserver.monitoring.MonitoringService;
 import org.maroubra.pemsserver.monitoring.MonitoringServiceImpl;
 
@@ -11,6 +13,7 @@ public class ServerBindings extends AbstractModule {
     @Override
     protected void configure() {
         bind(MonitoringService.class).to(MonitoringServiceImpl.class).in(Singleton.class);
+        bind(ControlService.class).to(ControlServiceImpl.class).in(Singleton.class);
         bind(MimetypesFileTypeMap.class).toInstance(new MimetypesFileTypeMap());
     }
 }
