--- pems-server/src/main/java/org/maroubra/pemsserver/configuration/ActuatorConfiguration.java
@@ -1,7 +1,0 @@
-package org.maroubra.pemsserver.configuration;
-
-import com.github.joschi.jadconfig.Parameter;
-
-public class ActuatorConfiguration {
-    // TODO
-}
