+++ pems-server/src/main/java/org/maroubra/pemsserver/configuration/Configuration.java
@@ -12,7 +12,6 @@ public class Configuration {
     private static ServerConfiguration serverConfiguration;
     private static MongoDbConfiguration mongoDbConfiguration;
     private static BluetoothConfiguration bluetoothConfiguration;
-    private static ActuatorConfiguration actuatorConfiguration;
 
     public static ServerConfiguration getServerConfiguration() {
         if (serverConfiguration == null)
@@ -35,18 +34,10 @@ public class Configuration {
         return bluetoothConfiguration;
     }
 
-    public static ActuatorConfiguration getActuatorConfiguration() {
-        if (actuatorConfiguration == null)
-            loadConfigurations();
-
-        return actuatorConfiguration;
-    }
-
     private static void loadConfigurations() {
         serverConfiguration = new ServerConfiguration();
         mongoDbConfiguration = new MongoDbConfiguration();
         bluetoothConfiguration = new BluetoothConfiguration();
-        actuatorConfiguration = new ActuatorConfiguration();
 
         JadConfig jadConfig = new JadConfig(new EnvironmentRepository(ENVIRONMENT_PREFIX), serverConfiguration, mongoDbConfiguration, bluetoothConfiguration);
         try {
