+++ pems-server/src/main/java/org/maroubra/pemsserver/configuration/MongoDbConfiguration.java
@@ -4,7 +4,7 @@ import com.github.joschi.jadconfig.Parameter;
 
 public class MongoDbConfiguration {
     @Parameter(value = "mongodb_connection")
-    public String connectionString = "mongodb://127.0.0.1:27017";
+    public String connectionString = "mongodb://0.0.0.0:27017";
 
     @Parameter(value = "mongodb_database")
     public String databaseName = "pems";
