+++ pems-server/src/main/java/org/maroubra/pemsserver/control/ControlService.java
@@ -1,0 +1,7 @@
+package org.maroubra.pemsserver.control;
+
+import java.util.List;
+
+public interface ControlService {
+    List<AbstractActuator> listActuators();
+}
